+++ scripts/build-example.sh
@@ -29,5 +29,9 @@ fi
   mkdir -p ./node_modules/algoliasearch-helper
   cp -R ../../node_modules/algoliasearch-helper/* ./node_modules/algoliasearch-helper
 
-  yarn build
+  if [[ "$name" != "server-side-rendering" ]]; then
+    yarn build
+  else
+    yarn build:ssr
+  fi
 )
